refactor: HotelEntity의 좌표 BigDecimal에서 Double로 변경

+++ src/main/java/staysplit/hotel_reservation/hotel/dto/request/CreateHotelRequest.java
@@ -1,12 +1,10 @@
 package staysplit.hotel_reservation.hotel.dto.request;
 
-import java.math.BigDecimal;
-
 public record CreateHotelRequest(
         String name,
         String address,
-        BigDecimal longitude,
-        BigDecimal latitude,
+        Double longitude,
+        Double latitude,
         String description,
         Integer starLevel
 ) {}
