+++ src/main/java/staysplit/hotel_reservation/hotel/dto/response/CreateHotelResponse.java
@@ -6,8 +6,8 @@ public record CreateHotelResponse(
         Integer hotelId,
         String name,
         String address,
-        BigDecimal longitude,
-        BigDecimal latitude,
+        Double longitude,
+        Double latitude,
         String description,
         Integer starLevel
 ) {
