+++ src/main/java/staysplit/hotel_reservation/hotel/dto/response/GetHotelDetailResponse.java
@@ -1,14 +1,13 @@
 package staysplit.hotel_reservation.hotel.dto.response;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 public record GetHotelDetailResponse(
         Integer hotelId,
         String name,
         String address,
-        BigDecimal longitude,
-        BigDecimal latitude,
+        Double longitude,
+        Double latitude,
         String description,
         Integer starLevel,
         Double rating,
